Add GetZoom function to set zoom values.
Zoom levels of plots can be set by editing the array at the
beginning of this function.
That's why the function should be at the top of the file, so that
it is easily accessible.

Plots obtain their desired zoom value by calling this function.

(this should be replaced by a GUI or something later)